_LibSyncer/requirements.txt
constants~=0.6.0
python-dotenv~=0.17.1
spotipy~=2.19.0
pandas~=1.4.1